Conversion of wolfSSL Manual to Markdown
The documentation has been converted from the Google Docs form to
Markdown to be compiled by pandoc (PDF) and mkdocs (HTML).

This allows documentation to be peer reviewed and tested as needed. It
has also allowed documentation cross-links and links to API calls for
references. The documentation has gone through a lot of cleanups as part
of the conversion process.

It still needs some work doing, such as an automated conversion of the
API documentation to be pulled into the required chapters and theming of
the HTML output.